+++ AndroidStudio/assignments/fundamentals/8th/AsyncTaskAssignment1/app/src/main/AndroidManifest.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="jp.mixi.assignment.async.beg"
-    android:versionCode="1"
-    android:versionName="1.0">
+    package="jp.mixi.assignment.async">
 
     <application
         android:allowBackup="true"
@@ -10,7 +8,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name="jp.mixi.assignment.async.beg.MainActivity"
+            android:name=".MainActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
