+++ AndroidStudio/assignments/fundamentals/8th/AsyncTaskAssignment1/app/src/main/java/jp/mixi/assignment/async/MainActivity.java
@@ -1,10 +1,10 @@
-package jp.mixi.assignment.async.beg;
+package jp.mixi.assignment.async;
 
-import android.app.Activity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
-public class MainActivity extends Activity {
+public class MainActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
